Update back-end service to save job alert email
