+++ src/main/java/com/techlooper/controller/UserController.java
@@ -29,147 +29,147 @@ import java.util.List;
 @RestController
 public class UserController {
 
-  @Resource
-  private ApplicationContext applicationContext;
+    @Resource
+    private ApplicationContext applicationContext;
 
-  @Resource
-  private UserService userService;
+    @Resource
+    private UserService userService;
 
-  @Resource
-  private TextEncryptor textEncryptor;
+    @Resource
+    private TextEncryptor textEncryptor;
 
-  @Resource
-  private UserEvaluationService userEvaluationService;
+    @Resource
+    private UserEvaluationService userEvaluationService;
 
-  @Resource
-  private PromotionService promotionService;
+    @Resource
+    private PromotionService promotionService;
 
-  @Resource
-  private CurrencyService currencyService;
+    @Resource
+    private CurrencyService currencyService;
 
-  @Resource
-  private Mapper dozerMapper;
+    @Resource
+    private Mapper dozerMapper;
 
-  @Resource
-  private SalaryReviewService salaryReviewService;
+    @Resource
+    private SalaryReviewService salaryReviewService;
 
-  @RequestMapping(value = "/api/users/add", method = RequestMethod.POST)
-  public void save(@RequestBody UserImportData userImportData, HttpServletResponse httpServletResponse) {
-    SocialProvider provider = userImportData.getCrawlerSource();
-    UserImportDataProcessor dataProcessor = applicationContext.getBean(provider + "UserImportDataProcessor", UserImportDataProcessor.class);
-    // process raw user data before import into ElasticSearch
-    List<UserImportEntity> userImportEntities = dataProcessor.process(Arrays.asList(userImportData));
-    httpServletResponse.setStatus(userService.addCrawledUser(userImportEntities.get(0), provider) ?
-      HttpServletResponse.SC_NO_CONTENT : HttpServletResponse.SC_NOT_ACCEPTABLE);
-  }
-
-  @RequestMapping(value = "/api/users/addAll", method = RequestMethod.POST)
-  public void saveAll(@RequestBody List<UserImportData> users, HttpServletResponse httpServletResponse) {
-    if (!users.isEmpty()) {
-      SocialProvider provider = users.get(0).getCrawlerSource();
-      UserImportDataProcessor dataProcessor = applicationContext.getBean(provider + "UserImportDataProcessor", UserImportDataProcessor.class);
-      // process raw user data before import into ElasticSearch
-      List<UserImportEntity> userImportEntities = dataProcessor.process(users);
-      httpServletResponse.setStatus(userService.addCrawledUserAll(userImportEntities, provider, UpdateModeEnum.MERGE) == users.size() ?
-        HttpServletResponse.SC_NO_CONTENT : HttpServletResponse.SC_NOT_ACCEPTABLE);
+    @RequestMapping(value = "/api/users/add", method = RequestMethod.POST)
+    public void save(@RequestBody UserImportData userImportData, HttpServletResponse httpServletResponse) {
+        SocialProvider provider = userImportData.getCrawlerSource();
+        UserImportDataProcessor dataProcessor = applicationContext.getBean(provider + "UserImportDataProcessor", UserImportDataProcessor.class);
+        // process raw user data before import into ElasticSearch
+        List<UserImportEntity> userImportEntities = dataProcessor.process(Arrays.asList(userImportData));
+        httpServletResponse.setStatus(userService.addCrawledUser(userImportEntities.get(0), provider) ?
+                HttpServletResponse.SC_NO_CONTENT : HttpServletResponse.SC_NOT_ACCEPTABLE);
     }
-    else {
-      httpServletResponse.setStatus(HttpServletResponse.SC_NOT_ACCEPTABLE);
+
+    @RequestMapping(value = "/api/users/addAll", method = RequestMethod.POST)
+    public void saveAll(@RequestBody List<UserImportData> users, HttpServletResponse httpServletResponse) {
+        if (!users.isEmpty()) {
+            SocialProvider provider = users.get(0).getCrawlerSource();
+            UserImportDataProcessor dataProcessor = applicationContext.getBean(provider + "UserImportDataProcessor", UserImportDataProcessor.class);
+            // process raw user data before import into ElasticSearch
+            List<UserImportEntity> userImportEntities = dataProcessor.process(users);
+            httpServletResponse.setStatus(userService.addCrawledUserAll(userImportEntities, provider, UpdateModeEnum.MERGE) == users.size() ?
+                    HttpServletResponse.SC_NO_CONTENT : HttpServletResponse.SC_NOT_ACCEPTABLE);
+        } else {
+            httpServletResponse.setStatus(HttpServletResponse.SC_NOT_ACCEPTABLE);
+        }
     }
-  }
 
-  @RequestMapping(value = "/user/save", method = RequestMethod.POST)
-  public List<FieldError> save(@RequestBody @Valid UserInfo userInfo, BindingResult result, HttpServletResponse httpServletResponse) {
-    if (result.getFieldErrorCount() > 0) {
-      httpServletResponse.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+    @RequestMapping(value = "/user/save", method = RequestMethod.POST)
+    public List<FieldError> save(@RequestBody @Valid UserInfo userInfo, BindingResult result, HttpServletResponse httpServletResponse) {
+        if (result.getFieldErrorCount() > 0) {
+            httpServletResponse.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+        } else {
+            userService.registerVietnamworksAccount(userInfo);
+            userService.save(userInfo);
+        }
+        return result.getFieldErrors();
     }
-    else {
-      userService.registerVietnamworksAccount(userInfo);
-      userService.save(userInfo);
+
+    @RequestMapping(value = "/api/user/findTalent", method = RequestMethod.POST)
+    public TalentSearchResponse findTalent(@RequestBody(required = false) TalentSearchRequest param, HttpServletResponse httpServletResponse) {
+        return userService.findTalent(param);
     }
-    return result.getFieldErrors();
-  }
 
-  @RequestMapping(value = "/api/user/findTalent", method = RequestMethod.POST)
-  public TalentSearchResponse findTalent(@RequestBody(required = false) TalentSearchRequest param, HttpServletResponse httpServletResponse) {
-    return userService.findTalent(param);
-  }
-
-  @RequestMapping(value = "/api/user/register", method = RequestMethod.POST)
-  public List<FieldError> registerUser(@RequestBody @Valid UserInfo userInfo, BindingResult result, HttpServletResponse httpServletResponse) {
-    if (result.getFieldErrorCount() > 0) {
-      httpServletResponse.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+    @RequestMapping(value = "/api/user/register", method = RequestMethod.POST)
+    public List<FieldError> registerUser(@RequestBody @Valid UserInfo userInfo, BindingResult result, HttpServletResponse httpServletResponse) {
+        if (result.getFieldErrorCount() > 0) {
+            httpServletResponse.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+        } else {
+            userService.registerUser(userInfo);
+        }
+        return result.getFieldErrors();
     }
-    else {
-      userService.registerUser(userInfo);
+
+    @SendToUser("/queue/info")
+    @MessageMapping("/user/findByKey")
+    @RequestMapping(value = "/user/findByKey", method = RequestMethod.POST)
+    public UserInfo getUserInfo(@CookieValue("techlooper.key") String techlooperKey/*, @DestinationVariable String username */) {
+        UserInfo userInfo = userService.findUserInfoByKey(techlooperKey);
+        userInfo.getLoginSource();
+        return userInfo;
     }
-    return result.getFieldErrors();
-  }
 
-  @SendToUser("/queue/info")
-  @MessageMapping("/user/findByKey")
-  @RequestMapping(value = "/user/findByKey", method = RequestMethod.POST)
-  public UserInfo getUserInfo(@CookieValue("techlooper.key") String techlooperKey/*, @DestinationVariable String username */) {
-    UserInfo userInfo = userService.findUserInfoByKey(techlooperKey);
-    userInfo.getLoginSource();
-    return userInfo;
-  }
-
-  @RequestMapping(value = "/user/verifyUserLogin", method = RequestMethod.POST)
-  public void verifyUserLogin(@RequestBody SocialRequest searchRequest, @CookieValue("techlooper.key") String techlooperKey, HttpServletResponse httpServletResponse) {
-    if (!textEncryptor.encrypt(searchRequest.getEmailAddress()).equals(techlooperKey)) {
-      httpServletResponse.setStatus(HttpServletResponse.SC_FORBIDDEN);
+    @RequestMapping(value = "/user/verifyUserLogin", method = RequestMethod.POST)
+    public void verifyUserLogin(@RequestBody SocialRequest searchRequest, @CookieValue("techlooper.key") String techlooperKey, HttpServletResponse httpServletResponse) {
+        if (!textEncryptor.encrypt(searchRequest.getEmailAddress()).equals(techlooperKey)) {
+            httpServletResponse.setStatus(HttpServletResponse.SC_FORBIDDEN);
+        }
     }
-  }
 
-  @RequestMapping(value = "/salaryReview", method = RequestMethod.POST)
-  public SalaryReviewDto reviewSalary(@RequestBody SalaryReview salaryReview) {
-    userEvaluationService.reviewSalary(salaryReview);
-    SalaryReviewDto salaryReviewDto = dozerMapper.map(salaryReview, SalaryReviewDto.class);
-    salaryReviewDto.setUsdToVndRate(currencyService.usdToVndRate());
-    return salaryReviewDto;
-  }
-
-  @RequestMapping(value = "/salaryReview/{id}", method = RequestMethod.GET)
-  public SalaryReviewDto getReviewSalary(@PathVariable("id") String id) {
-    return userService.findSalaryReviewById(id);
-  }
-
-  @RequestMapping(value = "/saveSalaryReviewSurvey", method = RequestMethod.POST)
-  public void saveSurvey(@RequestBody SalaryReviewSurvey salaryReviewSurvey, HttpServletResponse httpServletResponse) {
-    boolean isSaved = userEvaluationService.saveSalaryReviewSurvey(salaryReviewSurvey);
-    if (isSaved) {
-      httpServletResponse.setStatus(HttpServletResponse.SC_OK);
+    @RequestMapping(value = "/salaryReview", method = RequestMethod.POST)
+    public SalaryReviewDto reviewSalary(@RequestBody SalaryReview salaryReview) {
+        userEvaluationService.reviewSalary(salaryReview);
+        SalaryReviewDto salaryReviewDto = dozerMapper.map(salaryReview, SalaryReviewDto.class);
+        salaryReviewDto.setUsdToVndRate(currencyService.usdToVndRate());
+        return salaryReviewDto;
     }
-    else {
-      httpServletResponse.setStatus(HttpServletResponse.SC_METHOD_NOT_ALLOWED);
+
+    @RequestMapping(value = "/salaryReview/{id}", method = RequestMethod.GET)
+    public SalaryReviewDto getReviewSalary(@PathVariable("id") String id) {
+        return userService.findSalaryReviewById(id);
     }
-  }
 
-  @RequestMapping(value = "/priceJob", method = RequestMethod.POST)
-  public PriceJobEntity priceJob(@RequestBody PriceJobEntity priceJobEntity) {
-    userEvaluationService.priceJob(priceJobEntity);
-    return priceJobEntity;
-  }
-
-  @RequestMapping(value = "/savePriceJobSurvey", method = RequestMethod.POST)
-  public void savePriceJobSurvey(@RequestBody PriceJobSurvey priceJobSurvey, HttpServletResponse httpServletResponse) {
-    boolean isSaved = userEvaluationService.savePriceJobSurvey(priceJobSurvey);
-    if (isSaved) {
-      httpServletResponse.setStatus(HttpServletResponse.SC_OK);
+    @RequestMapping(value = "/saveSalaryReviewSurvey", method = RequestMethod.POST)
+    public void saveSurvey(@RequestBody SalaryReviewSurvey salaryReviewSurvey, HttpServletResponse httpServletResponse) {
+        boolean isSaved = userEvaluationService.saveSalaryReviewSurvey(salaryReviewSurvey);
+        if (isSaved) {
+            httpServletResponse.setStatus(HttpServletResponse.SC_OK);
+        } else {
+            httpServletResponse.setStatus(HttpServletResponse.SC_METHOD_NOT_ALLOWED);
+        }
     }
-    else {
-      httpServletResponse.setStatus(HttpServletResponse.SC_METHOD_NOT_ALLOWED);
+
+    @RequestMapping(value = "/priceJob", method = RequestMethod.POST)
+    public PriceJobEntity priceJob(@RequestBody PriceJobEntity priceJobEntity) {
+        userEvaluationService.priceJob(priceJobEntity);
+        return priceJobEntity;
     }
-  }
 
-  @RequestMapping("promotion/citibank/creditCard")
-  public void placeCitibankCreditCardPromotion(@Valid @RequestBody CitibankCreditCardPromotion citibankCreditCardPromotion) throws IOException, TemplateException {
-    promotionService.placePromotion(citibankCreditCardPromotion);
-  }
+    @RequestMapping(value = "/savePriceJobSurvey", method = RequestMethod.POST)
+    public void savePriceJobSurvey(@RequestBody PriceJobSurvey priceJobSurvey, HttpServletResponse httpServletResponse) {
+        boolean isSaved = userEvaluationService.savePriceJobSurvey(priceJobSurvey);
+        if (isSaved) {
+            httpServletResponse.setStatus(HttpServletResponse.SC_OK);
+        } else {
+            httpServletResponse.setStatus(HttpServletResponse.SC_METHOD_NOT_ALLOWED);
+        }
+    }
 
-  @RequestMapping(value = "salaryReview/placeSalaryReviewReport", method = RequestMethod.POST)
-  public void placeSalaryReviewReport(@Valid @RequestBody EmailRequest emailRequest) throws TemplateException, IOException, MessagingException {
-    salaryReviewService.sendSalaryReviewReportEmail(emailRequest);
-  }
+    @RequestMapping("promotion/citibank/creditCard")
+    public void placeCitibankCreditCardPromotion(@Valid @RequestBody CitibankCreditCardPromotion citibankCreditCardPromotion) throws IOException, TemplateException {
+        promotionService.placePromotion(citibankCreditCardPromotion);
+    }
+
+    @RequestMapping(value = "salaryReview/placeSalaryReviewReport", method = RequestMethod.POST)
+    public void placeSalaryReviewReport(@Valid @RequestBody EmailRequest emailRequest) throws TemplateException, IOException, MessagingException {
+        salaryReviewService.sendSalaryReviewReportEmail(emailRequest);
+    }
+
+    @MessageMapping("/jobs/createJobAlert")
+    public void createJobAlert(VnwJobAlertRequest vnwJobAlertRequest) {
+        salaryReviewService.createVnwJobAlert(vnwJobAlertRequest);
+    }
 }
